requirements.txt
# Core Deep Learning & RL
torch==2.0.1
numpy==1.24.3
gymnasium==0.28.1

# Web Framework
flask==2.3.2
flask-cors==4.0.0
marshmallow==3.19.0

# Background Jobs (Async Training)
celery==5.3.0
redis==4.5.5

# Configuration
pyyaml==6.0
python-dotenv==1.0.0

# Data & Visualization
matplotlib==3.7.1
pandas==2.0.2
seaborn==0.12.2
scipy==1.10.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Code Quality
black==23.3.0
flake8==6.0.0
isort==5.12.0
mypy==1.3.0

# Development
ipython==8.14.0
jupyter==1.0.0
ipykernel==6.24.0
